separacion de la logica del composable de las operaciones de la db
